+++ packages/create-faas-app/template/antd/server.ts
@@ -1,10 +1,17 @@
 import { dirname, join } from 'node:path'
+import { loadEnvFile } from 'node:process'
 import { fileURLToPath } from 'node:url'
 
 import { Server, staticHandler } from '@faasjs/core'
 
 const __filename = fileURLToPath(import.meta.url)
 const __dirname = dirname(__filename)
+
+try {
+  loadEnvFile()
+} catch (error) {
+  console.warn('[faasjs] Failed to load env file', error)
+}
 
 const publicHandler = staticHandler({
   root: join(__dirname, 'public'),